+++ Content_MultiAgent/src/seo_content_pipeline/models/__init__.py
@@ -21,6 +21,7 @@ from seo_content_pipeline.models.job import (
     StatusHistoryEntry,
 )
 from seo_content_pipeline.models.qa_result import QAReport
+from seo_content_pipeline.models.revision import RevisionHistoryArtifact, RevisionHistoryEntry
 from seo_content_pipeline.models.stage import StageView, WorkflowStage, WorkflowStatus
 from seo_content_pipeline.models.uniqueness import UniquenessProviderOption, UniquenessResult
 from seo_content_pipeline.models.validation import ValidationCheck, ValidationSeverity
@@ -39,6 +40,8 @@ __all__ = [
     "JobMetadata",
     "PipelineState",
     "QAReport",
+    "RevisionHistoryArtifact",
+    "RevisionHistoryEntry",
     "SEOBrief",
     "SEOBriefArtifact",
     "StageView",